feat: Extract CodeVerification to its own file
